feat: improve planner header
